Fix App AUMID conflict between H.NotifyIcon and ToastCOM
+ Also make ToastCOM's callback to work (for real this time).
